+++ CHANGELOG.md
@@ -13,8 +13,13 @@ The main contents of this release are the introduction of the project sharing an
     * Add granular access-control rules to `/api/v2` endpoints, valid for both project owners and guests (\#2999, \#3029).
     * Add pagination to `GET /admin/v2/task-group/activity/` and `GET /admin/v2/task-group/` (\#3023).
     * Do not cast endpoint return values to `PaginationResponse[X]` (\#3023).
+    * Reduce API logging level for some endpoints (\#3010).
     * Modify `GET /auth/current-user/allowed-viewer-paths/` logic, with `include_shared_projects` query parameter (\#3031).
     * Add validator for paths to forbid parent-directory references (\#3031).
+* App:
+    * Add `SlowResponseMiddleware` middleware (\#3035).
+* Settings:
+    * Add `Settings.FRACTAL_LONG_REQUEST_TIME` configuration variable (\#3035).
 * Database:
     * Add project-sharing-related `LinkUserProjectV2` columns (\#2999).
     * Move `UserOAuth.project_dir` to `.project_dirs` and drop `UserGrop.viewer_paths` (\#3031).